env: sunxi: Update default env fat device
Usually eMMC is default env fat device for environment,
if MMC_SUNXI_SLOT_EXTRA != 1 Sunxi always probed emmc
device as 1. but with DM_MMC it can be more possible to
probe eMMC as device 2 since for most of the sunxi platforms
eMMC is configured mmc2.

So update the env default device as 2 if DM_MMC and
MMC_SUNXI_SLOT_EXTRA != 1 but some boards may not use all
possible mmc devices or partly disabled in DT, for those update
the device in board specific defconfig.

Signed-off-by: Jagan Teki <jagan@amarulasolutions.com>